+++ stac_geoparquet/arrow/_api.py
@@ -4,9 +4,8 @@ from typing import Any, Dict, Iterable, Iterator, Optional, Union
 
 import pyarrow as pa
 
-from stac_geoparquet.arrow._batch import CleanBatch
+from stac_geoparquet.arrow._batch import CleanBatch, RawBatch
 from stac_geoparquet.arrow._schema.models import InferredSchema
-from stac_geoparquet.arrow._to_arrow import stac_items_to_arrow
 from stac_geoparquet.arrow._util import batched_iter
 from stac_geoparquet.json_reader import read_json_chunked
 
@@ -108,3 +107,28 @@ def stac_table_to_ndjson(table: pa.Table, dest: Union[str, os.PathLike[bytes]]) 
     for batch in table.to_batches():
         clean_batch = CleanBatch(batch)
         clean_batch.to_raw_batch().to_ndjson(dest)
+
+
+def stac_items_to_arrow(
+    items: Iterable[Dict[str, Any]], *, schema: Optional[pa.Schema] = None
+) -> pa.RecordBatch:
+    """Convert dicts representing STAC Items to Arrow
+
+    This converts GeoJSON geometries to WKB before Arrow conversion to allow multiple
+    geometry types.
+
+    All items will be parsed into a single RecordBatch, meaning that each internal array
+    is fully contiguous in memory for the length of `items`.
+
+    Args:
+        items: STAC Items to convert to Arrow
+
+    Kwargs:
+        schema: An optional schema that describes the format of the data. Note that this
+            must represent the geometry column as binary type.
+
+    Returns:
+        Arrow RecordBatch with items in Arrow
+    """
+    raw_batch = RawBatch.from_dicts(items, schema=schema)
+    return raw_batch.to_clean_batch().inner
