+++ stac_geoparquet/arrow/_to_arrow.py
@@ -1,40 +1,12 @@
 """Convert STAC data into Arrow tables"""
 
-from typing import Any, Dict, Iterable, Optional
-
 import ciso8601
 import numpy as np
 import orjson
 import pyarrow as pa
 import pyarrow.compute as pc
 
-from stac_geoparquet.arrow._batch import RawBatch
 from stac_geoparquet.arrow._crs import WGS84_CRS_JSON
-
-
-def stac_items_to_arrow(
-    items: Iterable[Dict[str, Any]], *, schema: Optional[pa.Schema] = None
-) -> pa.RecordBatch:
-    """Convert dicts representing STAC Items to Arrow
-
-    This converts GeoJSON geometries to WKB before Arrow conversion to allow multiple
-    geometry types.
-
-    All items will be parsed into a single RecordBatch, meaning that each internal array
-    is fully contiguous in memory for the length of `items`.
-
-    Args:
-        items: STAC Items to convert to Arrow
-
-    Kwargs:
-        schema: An optional schema that describes the format of the data. Note that this
-            must represent the geometry column as binary type.
-
-    Returns:
-        Arrow RecordBatch with items in Arrow
-    """
-    raw_batch = RawBatch.from_dicts(items, schema=schema)
-    return raw_batch.to_clean_batch().inner
 
 
 def bring_properties_to_top_level(
